+++ post-here-app/src/components/login/Login.jsx
@@ -14,15 +14,6 @@ const initialSignInFormErrors = {
   signInPassword: "",
 };
 
-const initialSignUpFormValues = {
-  signUpEmail: "",
-  signUpPassword: "",
-};
-const initialSignUpFormErrors = {
-  signUpEmail: "",
-  signUpPassword: "",
-  confirmPassword: "",
-};
 const initialDisabled = true;
 export default function Login() {
   const [signInFormValues, setSignInFormValues] = useState(
@@ -31,12 +22,7 @@ export default function Login() {
   const [signInFormErrors, setSignInFormErrors] = useState(
     initialSignInFormErrors
   );
-  const [signUpFormValues, setSignUpFormValues] = useState(
-    initialSignUpFormValues
-  );
-  const [signUpFormErrors, setSignUpFormErrors] = useState(
-    initialSignUpFormErrors
-  );
+
   const [disabled, setDisabled] = useState(initialDisabled);
 
   const signInOnInputChange = (evt) => {
@@ -67,27 +53,6 @@ export default function Login() {
     });
   }, [signInFormValues]);
 
-  const signUpOnInputChange = (evt) => {
-    const { name, value } = evt.target;
-    Yup.reach(signUpSchema, name)
-      .validate(value)
-      .then(() => {
-        setSignUpFormErrors({
-          ...signUpFormErrors,
-          [name]: "",
-        });
-      })
-      .catch((err) => {
-        setSignUpFormErrors({
-          ...signUpFormErrors,
-          [name]: err.errors[0],
-        });
-      });
-    setSignUpFormValues({
-      ...signUpFormValues,
-      [name]: value,
-    });
-  };
   const signInOnSubmit = (evt) => {
     evt.preventDefault();
 
@@ -97,31 +62,9 @@ export default function Login() {
     };
     console.log(signInUser);
   };
-  const signUpOnSubmit = (evt) => {
-    evt.preventDefault();
-
-    const signUpUser = {
-      email: signUpFormValues.signUpEmail.trim(),
-      password: signUpFormValues.signUpPassword.trim(),
-    };
-
-    console.log(signUpUser);
-  };
-  useEffect(() => {
-    signUpSchema.isValid(signUpFormValues).then((valid) => {
-      setDisabled(!valid);
-    });
-  }, [signUpFormValues]);
 
   return (
     <div>
-      <SignUpForm
-        values={signInFormValues}
-        onSubmit={signUpOnSubmit}
-        onInputChange={signUpOnInputChange}
-        disabled={disabled}
-        formErrors={signUpFormErrors}
-      />
       <SignInForm
         values={signInFormValues}
         onSubmit={signInOnSubmit}
