+++ post-here-app/src/components/login/SignUp.jsx
@@ -1,0 +1,74 @@
+import React, { useState, useEffect } from "react";
+import SignUpForm from "./loginForms/SignUpForm.jsx";
+import signUpSchema from "./validation/signUpSchema.js";
+import * as Yup from "yup";
+
+const initialSignUpFormValues = {
+  signUpEmail: "",
+  signUpPassword: "",
+};
+const initialSignUpFormErrors = {
+  signUpEmail: "",
+  signUpPassword: "",
+  confirmPassword: "",
+};
+const initialDisabled = true;
+function SignUp() {
+  const [signUpFormValues, setSignUpFormValues] = useState(
+    initialSignUpFormValues
+  );
+  const [signUpFormErrors, setSignUpFormErrors] = useState(
+    initialSignUpFormErrors
+  );
+  const [disabled, setDisabled] = useState(initialDisabled);
+
+  const signUpOnInputChange = (evt) => {
+    const { name, value } = evt.target;
+    Yup.reach(signUpSchema, name)
+      .validate(value)
+      .then(() => {
+        setSignUpFormErrors({
+          ...signUpFormErrors,
+          [name]: "",
+        });
+      })
+      .catch((err) => {
+        setSignUpFormErrors({
+          ...signUpFormErrors,
+          [name]: err.errors[0],
+        });
+      });
+    setSignUpFormValues({
+      ...signUpFormValues,
+      [name]: value,
+    });
+  };
+  const signUpOnSubmit = (evt) => {
+    evt.preventDefault();
+
+    const signUpUser = {
+      email: signUpFormValues.signUpEmail.trim(),
+      password: signUpFormValues.signUpPassword.trim(),
+    };
+
+    console.log(signUpUser);
+  };
+  useEffect(() => {
+    signUpSchema.isValid(signUpFormValues).then((valid) => {
+      setDisabled(!valid);
+    });
+  }, [signUpFormValues]);
+  return (
+    <div>
+      <SignUpForm
+        values={signInFormValues}
+        onSubmit={signUpOnSubmit}
+        onInputChange={signUpOnInputChange}
+        disabled={disabled}
+        formErrors={signUpFormErrors}
+      />
+    </div>
+  );
+}
+
+export default SignUp;
